% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_raster_square.R
\name{crop_raster_square}
\alias{crop_raster_square}
\title{Crops a raster and returns a smaller square raster}
\usage{
crop_raster_square(raster_input, lat, long, square_km,
  increase_resolution = 1)
}
\arguments{
\item{raster_input}{a raster}

\item{lat}{decimal latitude of the centre of the cropped square}

\item{long}{decimal longitude of the centre of the cropped square}

\item{square_km}{length of one side of the square in km}

\item{increase_resolution}{optional multiplier to increase number of cells in the raster. Default = 1.}
}
\value{
cropped raster
}
\description{
Crops a raster and returns a smaller square raster
}
\examples{
crop_raster_square(a_raster, 54.4282623, -2.9787427, square_km = 10)
}
