% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_raster_track.R
\name{crop_raster_track}
\alias{crop_raster_track}
\title{Crops a raster into a rectangle surrounding a set of lat long points}
\usage{
crop_raster_track(raster_input, lat_points, long_points,
  width_buffer = 1, increase_resolution = 1)
}
\arguments{
\item{raster_input}{a raster}

\item{lat_points}{a vector of decimal latitudes}

\item{long_points}{a vector of decimal longitudes}

\item{increase_resolution}{optional multiplier to increase number of cells in the raster. Default = 1.}

\item{width_bufer}{buffer distance around the provided points}
}
\value{
cropped raster
}
\description{
Crops a raster into a rectangle surrounding a set of lat long points
}
\examples{
crop_raster_track(a_raster, some_lat_points, some_long_points)
}
