% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_shade.R
\name{elevation_shade}
\alias{elevation_shade}
\title{Produces an elevation shaded png from a raster}
\usage{
elevation_shade(raster_dem, elevation_palette = c("#54843f", "#808080",
  "#FFFFFF"), return_png = TRUE, png_opacity = 1)
}
\arguments{
\item{raster_dem}{a raster}

\item{elevation_palette}{a vector of colours to use for elevation shading}

\item{return_png}{\code{TRUE} to return a png image. \code{FALSE} will return a raster}

\item{png_opacity}{Opacity of the returned image if requesting a png}
}
\value{
elevation shaded png image
}
\description{
Produces an elevation shaded png from a raster
}
\examples{
elevation_shade(example_raster())
}
