% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extdata.R
\name{example_igc}
\alias{example_igc}
\title{Returns an example IGC file using read_igc()}
\usage{
example_igc()
}
\value{
a tibble
}
\description{
Returns an example IGC file using read_igc()
}
\examples{

# Loads a paragliding flight GPS track, originally downloaded from xcleague.com

igc <- example_igc()
}
