% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extdata.R
\name{example_raster}
\alias{example_raster}
\title{Returns an example digital elevation model raster file()}
\usage{
example_raster()
}
\value{
a raster
}
\description{
Returns an example digital elevation model raster file()
}
\examples{
example_raster <- example_raster()
}
