% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_image.R
\name{ggmap_image}
\alias{ggmap_image}
\title{Gets a ggmap to overlay a rayshader scene and returns a raster}
\usage{
ggmap_image(raster_input, ...)
}
\arguments{
\item{raster_input}{a raster to use for the dimensions of the overlay}

\item{...}{Additional parameters to pass to ggmap::get_map() e.g. maptype, source, zoom, color}
}
\value{
a png image to overlay a rayshader scene
}
\description{
Gets a ggmap to overlay a rayshader scene and returns a raster
}
\examples{
ggmap_image(a_raster, maptype = "Stamen", zoom = 11, color = "bw")

}
