% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapzen_dem.R
\name{mapzen_dem}
\alias{mapzen_dem}
\title{Gets Digital Elevation Model (DEM) data from 'mapzen' via 'Amazon Public Datasets'}
\usage{
mapzen_dem(lat, long, square_km, api_key, max_tiles = 10)
}
\arguments{
\item{lat}{WGS84 latitude}

\item{long}{WGS84 longitude}

\item{square_km}{length of one edge the required square area, in km}
}
\value{
a raster with values corresponding to terrain height in metres
}
\description{
Gets Digital Elevation Model (DEM) data from 'mapzen' via 'Amazon Public Datasets'
}
\examples{
lat = 54.4502651
long = -3.1767946
square_km = 10

dem <- mapzen_dem(lat, long, square_km)
}
