% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_asc.R
\name{mosaic_asc}
\alias{mosaic_asc}
\title{Stitches together .asc files into a single raster
Requires a target directory of .asc files or .zip files containing .asc files}
\usage{
mosaic_asc(asc_path, extract_zip = FALSE, file_match = ".*.asc",
  zip_file_match = ".*.zip", raster_output_file = "mosaic_out.raster")
}
\arguments{
\item{asc_path}{path to files that are to be stitched together}

\item{extract_zip}{\code{FALSE} to target .asc files, \code{TRUE} if your .asc files are zipped.}

\item{file_match}{regex pattern to match .asc files, either in asc_path or in zip files.}

\item{zip_file_match}{regex pattern to match .zip files}

\item{raster_output_file}{raster file to be created (will overwrite existing files)}
}
\value{
TRUE
}
\description{
Stitches together .asc files into a single raster
Requires a target directory of .asc files or .zip files containing .asc files
}
\examples{
\donttest{
mosaic_asc("path/to/grid/zip_files/", "output/file.raster")
}
}
