% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_ASTER.R
\name{mosaic_ASTER}
\alias{mosaic_ASTER}
\title{Stitches together ASTER GDEM TIF files and saves as a single large raster
Requires a target directory of ASTER zip files. Effecively a duplicate of mosaic_uk_grid() with a few tweaks.}
\usage{
mosaic_ASTER(aster_path, raster_output_file = "mosaic_ASTER.raster")
}
\arguments{
\item{aster_path}{directory containing OS Terrain 50 'GRID' files}

\item{raster_output_file}{path and filename for the merged raster. End in ".raster" to save in default R Raster format (.grd)}
}
\value{
TRUE
}
\description{
Stitches together ASTER GDEM TIF files and saves as a single large raster
Requires a target directory of ASTER zip files. Effecively a duplicate of mosaic_uk_grid() with a few tweaks.
}
\examples{
mosaic_ASTER("path/to/grid/zip_files/", "output/file.raster")
}
