% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_defra_lidar.R
\name{mosaic_defra_lidar}
\alias{mosaic_defra_lidar}
\title{Stitches together DEFRA Lidar files and saves as a single large raster
Requires a target directory of LIDAR zip files.}
\usage{
mosaic_defra_lidar(lidar_path,
  raster_output_file = "mosaic_uk_grid.raster")
}
\arguments{
\item{lidar_path}{directory containing defra lidar zip files}

\item{raster_output_file}{path and filename for the merged raster. End in ".raster" to save in default R Raster format (.grd)}
}
\value{
TRUE
}
\description{
Stitches together DEFRA Lidar files and saves as a single large raster
Requires a target directory of LIDAR zip files.
}
\examples{
\donttest{
mosaic_defra_lidar("path/to/grid/zip_files/", "output/file.raster")
}
}
