% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_uk_grid.R
\name{mosaic_uk_grid}
\alias{mosaic_uk_grid}
\title{Stitches together OS Terrain 50 'GRID' files and saves as a single large raster
Requires a target directory of 'GRID' zip files.}
\usage{
mosaic_uk_grid(os50_path, raster_output_file = "mosaic_uk_grid.raster")
}
\arguments{
\item{os50_path}{directory containing OS Terrain 50 'GRID' files}

\item{raster_output_file}{path and filename for the merged raster. End in ".raster" to save in default R Raster format (.grd)}
}
\value{
TRUE
}
\description{
Stitches together OS Terrain 50 'GRID' files and saves as a single large raster
Requires a target directory of 'GRID' zip files.
}
\examples{
\donttest{
mosaic_uk_grid("path/to/grid/zip_files/", "output/file.raster")
}
}
