% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_igc.R
\name{read_igc}
\alias{read_igc}
\title{Load an IGC file}
\usage{
read_igc(path)
}
\arguments{
\item{path}{target IGC file}
}
\value{
processed IGC file
}
\description{
Load an IGC file
}
\examples{
\donttest{
read_igc("file.igc")
}
}
