% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippy_raster.R
\name{slippy_raster}
\alias{slippy_raster}
\title{Creates an overlay image from various sources using Miles McBain's slippymath}
\usage{
slippy_raster(lat, long, square_km, image_source = "stamen",
  image_type = "watercolor", resolution = 1000, max_tiles = 30,
  api_key)
}
\arguments{
\item{lat}{WGS84 latitude}

\item{long}{WGS84 longitude}

\item{square_km}{length of one edge the required square area, in km}

\item{image_source}{Source for the overlay image. Valid entries are "mapbox", "stamen".}

\item{image_type}{The type of overlay to request. "satellite", "mapbox-streets-v8", "mapbox-terrain-v2", "mapbox-traffic-v1", "terrain-rgb", "mapbox-incidents-v1" (mapbox) or "watercolor", "toner", "terrain" (stamen)}

\item{resolution}{width and height (cell count) of the returned raster}

\item{max_tiles}{Maximum number of tiles to be requested by slippymath}

\item{api_key}{API key (required for mapbox)}
}
\value{
a rasterBrick image
}
\description{
Creates an overlay image from various sources using Miles McBain's slippymath
}
\examples{
\donttest{
#NEEDS EXAMPLE
}
}
