% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gps_to_rayshader.R
\name{add_gps_to_rayshader}
\alias{add_gps_to_rayshader}
\title{Adds a GPS trace to a rayshader scene}
\usage{
add_gps_to_rayshader(raster_input, lat, long, alt, zscale,
  line_width = 1.5, colour = "red", alpha = 0.8)
}
\arguments{
\item{raster_input}{a raster}

\item{lat}{vector of decimal latitude points}

\item{long}{vector of decimal longitude points}

\item{alt}{vector of altitudes}
}
\value{
Adds GPS trace to the current rayshader scene
}
\description{
Adds a GPS trace to a rayshader scene
}
\examples{
flight <- read_igc("path/to/igc/file")
add_gps_to_rayshader(a_raster, flight$lat, flight$long, flight$altitude)
}
