% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_raster_square.R
\name{crop_raster_square}
\alias{crop_raster_square}
\title{Crops a raster and returns a smaller square raster}
\usage{
crop_raster_square(rasterIn, lat, long, square_km, increase_resolution = 1)
}
\arguments{
\item{rasterIn}{a raster}

\item{lat}{WGS84 latitude of the centre of the cropped square}

\item{long}{WGS84 longitude of the centre of the cropped square}

\item{square_km}{length of one side of the square in km}

\item{increase_resolution}{optional multiplier to increase number of cells in the raster}
}
\value{
A cropped raster
}
\description{
Crops a raster and returns a smaller square raster
}
\examples{
crop_raster_square(example_raster(), lat = 54.513293, long = -3.045598, square_km = 0.01)
}
