% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_shade.R
\name{elevation_shade}
\alias{elevation_shade}
\title{Produces an elevation shaded png from a raster}
\usage{
elevation_shade(raster_input, elevation_palette = c("#54843f", "#808080",
  "#FFFFFF"))
}
\arguments{
\item{raster_input}{a raster}

\item{elevation_palette}{a vector of colours to use for elevation shading}
}
\value{
elevation shaded png image
}
\description{
Produces an elevation shaded png from a raster
}
\examples{
elevation_shade(example_raster)
}
