% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_raster}
\alias{example_raster}
\title{A small elevation raster}
\format{A RasterLayer 645x645
\describe{
  \item{values}{terrain elevation in metres}

}}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1.1?tab=mapview}
}
\usage{
example_raster
}
\description{
25m resolution Digital Elevation Map of part of the UK Lake District
}
\keyword{datasets}
