% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extdata.R
\name{example_raster}
\alias{example_raster}
\title{Returns an example digital elevation model raster file()}
\usage{
example_raster()
}
\value{
a raster
}
\description{
Returns an example digital elevation model raster file()
}
\examples{

# Load elevation data describing  a small section of the English Lake District
# Source: EU Copernicus https://land.copernicus.eu/terms-of-use

example_raster <- example_raster()
}
