% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggslippy.R
\name{ggslippy}
\alias{ggslippy}
\title{Creates an overlay image from various sources using Miles McBain's slippymath}
\usage{
ggslippy(slippy_raster)
}
\arguments{
\item{slippy_raster}{A raster raster returned by either \code{slippy_raster()} or \code{slippy_overlay(return_png = FALSE)}}
}
\value{
a ggplot object
}
\description{
Creates an overlay image from various sources using Miles McBain's slippymath
}
\examples{
\donttest{
example...
}
}
