% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_files.R
\name{mosaic_files}
\alias{mosaic_files}
\title{Stitches together files into a single raster
Requires a target directory of files that can be read with raster::raster(), e.g. .asc files, or a directory of .zip files containing these files}
\usage{
mosaic_files(path, extract_zip = FALSE, file_match = ".*.asc",
  zip_file_match = ".*.zip", raster_output_file = "mosaic_out.raster",
  file_crs = NULL, raster_todisk = FALSE)
}
\arguments{
\item{path}{path to files that are to be stitched together}

\item{extract_zip}{\code{FALSE} to target .asc files, \code{TRUE} if your .asc files are zipped.}

\item{file_match}{regex pattern to match .asc files, either in \code{path} or in zip files.}

\item{zip_file_match}{regex pattern to match .zip files}

\item{raster_output_file}{raster file to be created (will overwrite existing files)}

\item{file_crs}{projection string of the input files. Output will always be WGS84.}

\item{raster_todisk}{Setting \code{TRUE} will set \code{rasterOptions(todisk=TRUE)}, which can help with memory issues.}
}
\value{
TRUE
}
\description{
Stitches together files into a single raster
Requires a target directory of files that can be read with raster::raster(), e.g. .asc files, or a directory of .zip files containing these files
}
\examples{
\dontrun{
mosaic_files("path/to/grid/zip_files/", "output/file.raster")

raster_mosaic <- raster::raster("output/file.gri")
}
}
