% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_zscale.R
\name{raster_zscale}
\alias{raster_zscale}
\title{Approximates the zscale of a raster Digital Elevation Model for 'rayshader'}
\usage{
raster_zscale(raster, height_units = "m")
}
\arguments{
\item{raster}{A raster object of elevation data values}

\item{height_units}{Elevation units of the raster, c("m", "feet")}
}
\value{
a number to be used as zscale in rayshader::plot_3d()
}
\description{
Approximates the zscale of a raster Digital Elevation Model for 'rayshader'
}
\examples{
raster_zscale(example_raster())

}
