% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_location.R
\name{render_location}
\alias{render_location}
\title{Colours and hill shades a raster using rayshader}
\usage{
render_location(raster_input, sphere_palette = "bw", sunangle = 270,
  fast = FALSE, add_water = TRUE, water_min_area = 20,
  water_cutoff = 0.995, water_colour = "imhof4", img_layers = NULL)
}
\arguments{
\item{raster_input}{a raster}

\item{sphere_palette}{A rayshader palette to use for sphere_shade(), e.g. "bw" or "desert".
This palette will be multiplied by your elevation palette so use "bw" to maintain original hues.}

\item{sunangle}{rayshader sun angle}

\item{fast}{skips ray shade and ambient shade to render quickly for checking}

\item{add_water}{}

\item{elevation_palette}{a vector of colours to use for elevation shading}
}
\value{
rendedered scene to visualise with rayshader::plot_3d()
returns a list. r$r is the rendered scene and r$elmat is the elevation matrix
}
\description{
crop large rasters first using crop_raster_square()
Use rayshader::plot_3d() to visualise the result
}
\examples{
render_location(a_raster, 54.4282623, -2.9787427, square_km = 10)

plot_3d(
  r$r,
  r$elmat,
  zscale = 50,
  shadow = TRUE
)
}
